requirements.txt
opencv-python
pushbullet.py